The functionallity of all the functions were finished
feat: All functions have the functionalities ready for testing. Two parameters were added in two functions.
doc: Comments were added on some parts of the code.

Comments were added on some parts of the code that was considered to be required in order make the code easier to understand.
The functions have their functionalities finished ready for testing. The parameters were added at the BidList_End() and at the BidList_Empty() functions
in order to enable this functions to run