Genereal bug fixes
Feat: A new makefile was created.
Fix: Bugs were fixed in the .h file and in the main.c files

A new makefile was created to look more professional. And endif
was added in the .h file in order solve the warning showed previously when compiled.
A '!' was added in the condition of the while in order to allow the pointer to travel to the end of the list.

The main file, now proves the funtionality of the list.