Bug fixed on the create function and a better demonstration in the main function. The list is now finished
fix: A bug was solved in the create function. It didn't allow a node to point at the previous node.
feat: We can see in the main.c that all the functions in the BidList library are called in order to prove its functionality.

The Bidirectional linked list is now considered finished. In the main.c we can see its functionality and we prove that it works